requirements.txt
fastapi
pymongo
python-dotenv
openai
langchain
langchain_openai
langchain_community
pytest
uvicorn
